[refactor] Moved base metrics class
