+++ .gitignore
@@ -1,5 +1,6 @@
 __pycache__
 /dist
+/.venv
 /.vscode
 /.pytest_*
 **.egg-info
