+++ src/mlinstruct/evaluate/metrics/__init__.py
@@ -1,4 +1,5 @@
-from . import classification
+from .base_metrics import BaseMetrics
 from .metric_utils import MetricUtils
+from . import classification
 
-__all__ = ["classification", "MetricUtils"]
+__all__ = ["BaseMetrics", "MetricUtils", "classification"]
