+++ src/mlinstruct/evaluate/metrics/classification/__init__.py
@@ -1,5 +1,4 @@
-from .base_metrics import BaseMetrics
 from .confusion_matrix import ConfusionMatrix
 from .roc import ROC
 
-__all__ = ["BaseMetrics", "ConfusionMatrix", "ROC"]
+__all__ = ["ConfusionMatrix", "ROC"]
